ND: bills scraper rewrite, now has sponsors and random sleeps
